+++ avk/algorithms/selection_sort.cpp
@@ -2,10 +2,10 @@
 
 void selection_sort(main_array& array)
 {
-	for (int i = 0; i < array.size() - 1; ++i)
+	for (int i = 0; i < (int)array.size() - 1; ++i)
 	{
 		int min = i;
-		for (int j = i + 1; j < array.size(); ++j)
+		for (int j = i + 1; j < (int)array.size(); ++j)
 			if (array[j] < array[min])
 				min = j;
 		if (min != i)
